requirements.txt
asgiref==3.5.2
Django==4.1
django-extensions==3.2.0
djangorestframework==3.13.1
mysql-client==0.0.1
PyMySQL==1.0.2
pytz==2022.2.1
sqlparse==0.4.2
cryptography==37.0.2
gunicorn
watchfiles
django-clone==3.0.6